+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- `Innmind\Reflection\ReflectionType::type(): Innmind\Type\Type`
+
 ### Removed
 
 - Support for PHP `8.1`